[Rules migration] Post merge feedback followups (#202815)
## Summary

These are the followup updated to address feedback from my previous PRs:

* Make sure to use descriptive names specific to the `siem_migrations`
subdomain
([comment](https://github.com/elastic/kibana/pull/200978#pullrequestreview-2454582368)):

> Make sure you use descriptive names specific to the siem_migrations
subdomain. Names like RulesPage, RulesTable, useRulesColumns etc are way
too generic and conflict with the rule management terminology, which
would make code search more difficult.

* Export the memo component directly everywhere
([comment](https://github.com/elastic/kibana/pull/201597#discussion_r1858069127)):

> Could we export the memo component directly everywhere? It's shorter
and it makes it easier to find the references in the IDE.

* Use one hook to access APIs instead of two
([comment](https://github.com/elastic/kibana/pull/202494#discussion_r1867967135)):

> I see that for every API request we have to implement 2 separate
hooks. Why don't we add error handling to the same hook that does the
useQuery? so we have everything in one hook. Or is there a reason to
have them separate?

(cherry picked from commit 4d8f7111d0f8c330aab2c8347cf6d131570ff665)
