+++ x-pack/plugins/security_solution/public/siem_migrations/routes.tsx
@@ -10,16 +10,19 @@ import { Routes, Route } from '@kbn/shared-ux-router';
 
 import type { SecuritySubPluginRoutes } from '../app/types';
 import { SIEM_MIGRATIONS_RULES_PATH, SecurityPageName } from '../../common/constants';
-import { RulesPage } from './rules/pages';
+import { MigrationRulesPage } from './rules/pages';
 import { PluginTemplateWrapper } from '../common/components/plugin_template_wrapper';
 import { SecurityRoutePageWrapper } from '../common/components/security_route_page_wrapper';
 
-export const RulesRoutes = () => {
+export const SiemMigrationsRoutes = () => {
   return (
     <PluginTemplateWrapper>
       <SecurityRoutePageWrapper pageName={SecurityPageName.siemMigrationsRules}>
         <Routes>
-          <Route path={`${SIEM_MIGRATIONS_RULES_PATH}/:migrationId?`} component={RulesPage} />
+          <Route
+            path={`${SIEM_MIGRATIONS_RULES_PATH}/:migrationId?`}
+            component={MigrationRulesPage}
+          />
         </Routes>
       </SecurityRoutePageWrapper>
     </PluginTemplateWrapper>
@@ -29,6 +32,6 @@ export const RulesRoutes = () => {
 export const routes: SecuritySubPluginRoutes = [
   {
     path: SIEM_MIGRATIONS_RULES_PATH,
-    component: RulesRoutes,
+    component: SiemMigrationsRoutes,
   },
 ];
