+++ x-pack/plugins/security_solution/public/siem_migrations/rules/components/rule_details_flyout/translations.ts
