+++ x-pack/plugins/security_solution/public/siem_migrations/rules/components/rules_table/index.tsx
@@ -17,20 +17,22 @@ import {
 } from '@elastic/eui';
 import React, { useCallback, useMemo, useState } from 'react';
 
+import { useAppToasts } from '../../../../common/hooks/use_app_toasts';
 import type { RuleMigration } from '../../../../../common/siem_migrations/model/rule_migration.gen';
-import { NoItemsMessage } from './no_items_message';
-import { useRulesTableColumns } from '../../hooks/use_rules_table_columns';
-import { useRulePreviewFlyout } from '../../hooks/use_rule_preview_flyout';
+import { EmptyMigration } from './empty_migration';
+import { useMigrationRulesTableColumns } from '../../hooks/use_migration_rules_table_columns';
+import { useMigrationRuleDetailsFlyout } from '../../hooks/use_migration_rule_preview_flyout';
 import { useInstallMigrationRules } from '../../logic/use_install_migration_rules';
 import { useGetMigrationRules } from '../../logic/use_get_migration_rules';
 import { useInstallTranslatedMigrationRules } from '../../logic/use_install_translated_migration_rules';
-import { BulkActions } from './bulk_actions';
 import { useGetMigrationTranslationStats } from '../../logic/use_get_migration_translation_stats';
+import * as logicI18n from '../../logic/translations';
+import { BulkActions } from './bulk_actions';
 import { SearchField } from './search_field';
 
 const DEFAULT_PAGE_SIZE = 10;
 
-export interface RulesTableComponentProps {
+export interface MigrationRulesTableProps {
   /**
    * Selected rule migration id
    */
@@ -40,143 +42,152 @@ export interface RulesTableComponentProps {
 /**
  * Table Component for displaying SIEM rules migrations
  */
-const RulesTableComponent: React.FC<RulesTableComponentProps> = ({ migrationId }) => {
-  const [pageIndex, setPageIndex] = useState(0);
-  const [pageSize, setPageSize] = useState(DEFAULT_PAGE_SIZE);
-  const [searchTerm, setSearchTerm] = useState<string | undefined>();
+export const MigrationRulesTable: React.FC<MigrationRulesTableProps> = React.memo(
+  ({ migrationId }) => {
+    const { addError } = useAppToasts();
 
-  const { data: translationStats, isLoading: isStatsLoading } =
-    useGetMigrationTranslationStats(migrationId);
+    const [pageIndex, setPageIndex] = useState(0);
+    const [pageSize, setPageSize] = useState(DEFAULT_PAGE_SIZE);
+    const [searchTerm, setSearchTerm] = useState<string | undefined>();
 
-  const {
-    data: { ruleMigrations, total } = { ruleMigrations: [], total: 0 },
-    isLoading: isDataLoading,
-  } = useGetMigrationRules({
-    migrationId,
-    page: pageIndex,
-    perPage: pageSize,
-    searchTerm,
-  });
+    const { data: translationStats, isLoading: isStatsLoading } =
+      useGetMigrationTranslationStats(migrationId);
 
-  const [selectedRuleMigrations, setSelectedRuleMigrations] = useState<RuleMigration[]>([]);
+    const {
+      data: { ruleMigrations, total } = { ruleMigrations: [], total: 0 },
+      isLoading: isDataLoading,
+    } = useGetMigrationRules({
+      migrationId,
+      page: pageIndex,
+      perPage: pageSize,
+      searchTerm,
+    });
 
-  const pagination = useMemo(() => {
-    return {
-      pageIndex,
-      pageSize,
-      totalItemCount: total,
-    };
-  }, [pageIndex, pageSize, total]);
+    const [selectedRuleMigrations, setSelectedRuleMigrations] = useState<RuleMigration[]>([]);
 
-  const onTableChange = useCallback(({ page, sort }: CriteriaWithPagination<RuleMigration>) => {
-    if (page) {
-      setPageIndex(page.index);
-      setPageSize(page.size);
-    }
-  }, []);
+    const pagination = useMemo(() => {
+      return {
+        pageIndex,
+        pageSize,
+        totalItemCount: total,
+      };
+    }, [pageIndex, pageSize, total]);
 
-  const handleOnSearch = useCallback((value: string) => {
-    setSearchTerm(value.trim());
-  }, []);
-
-  const { mutateAsync: installMigrationRules } = useInstallMigrationRules(migrationId);
-  const { mutateAsync: installTranslatedMigrationRules } =
-    useInstallTranslatedMigrationRules(migrationId);
-
-  const [isTableLoading, setTableLoading] = useState(false);
-  const installSingleRule = useCallback(
-    async (migrationRule: RuleMigration, enable?: boolean) => {
-      setTableLoading(true);
-      try {
-        await installMigrationRules([migrationRule.id]);
-      } finally {
-        setTableLoading(false);
+    const onTableChange = useCallback(({ page, sort }: CriteriaWithPagination<RuleMigration>) => {
+      if (page) {
+        setPageIndex(page.index);
+        setPageSize(page.size);
       }
-    },
-    [installMigrationRules]
-  );
+    }, []);
 
-  const installTranslatedRules = useCallback(
-    async (enable?: boolean) => {
-      setTableLoading(true);
-      try {
-        await installTranslatedMigrationRules();
-      } finally {
-        setTableLoading(false);
-      }
-    },
-    [installTranslatedMigrationRules]
-  );
+    const handleOnSearch = useCallback((value: string) => {
+      setSearchTerm(value.trim());
+    }, []);
 
-  const ruleActionsFactory = useCallback(
-    (ruleMigration: RuleMigration, closeRulePreview: () => void) => {
-      // TODO: Add flyout action buttons
-      return null;
-    },
-    []
-  );
+    const { mutateAsync: installMigrationRules } = useInstallMigrationRules(migrationId);
+    const { mutateAsync: installTranslatedMigrationRules } =
+      useInstallTranslatedMigrationRules(migrationId);
 
-  const { rulePreviewFlyout, openRulePreview } = useRulePreviewFlyout({
-    ruleActionsFactory,
-  });
-
-  const rulesColumns = useRulesTableColumns({
-    disableActions: isTableLoading,
-    openMigrationRulePreview: openRulePreview,
-    installMigrationRule: installSingleRule,
-  });
-
-  return (
-    <>
-      <EuiSkeletonLoading
-        isLoading={isDataLoading}
-        loadingContent={
-          <>
-            <EuiSkeletonTitle />
-            <EuiSkeletonText />
-          </>
+    const [isTableLoading, setTableLoading] = useState(false);
+    const installSingleRule = useCallback(
+      async (migrationRule: RuleMigration, enable?: boolean) => {
+        setTableLoading(true);
+        try {
+          await installMigrationRules([migrationRule.id]);
+        } catch (error) {
+          addError(error, { title: logicI18n.INSTALL_MIGRATION_RULES_FAILURE });
+        } finally {
+          setTableLoading(false);
         }
-        loadedContent={
-          !translationStats?.rules.total ? (
-            <NoItemsMessage />
-          ) : (
+      },
+      [addError, installMigrationRules]
+    );
+
+    const installTranslatedRules = useCallback(
+      async (enable?: boolean) => {
+        setTableLoading(true);
+        try {
+          await installTranslatedMigrationRules();
+        } catch (error) {
+          addError(error, { title: logicI18n.INSTALL_MIGRATION_RULES_FAILURE });
+        } finally {
+          setTableLoading(false);
+        }
+      },
+      [addError, installTranslatedMigrationRules]
+    );
+
+    const ruleActionsFactory = useCallback(
+      (ruleMigration: RuleMigration, closeRulePreview: () => void) => {
+        // TODO: Add flyout action buttons
+        return null;
+      },
+      []
+    );
+
+    const {
+      migrationRuleDetailsFlyout: rulePreviewFlyout,
+      openMigrationRuleDetails: openRulePreview,
+    } = useMigrationRuleDetailsFlyout({
+      ruleActionsFactory,
+    });
+
+    const rulesColumns = useMigrationRulesTableColumns({
+      disableActions: isTableLoading,
+      openMigrationRuleDetails: openRulePreview,
+      installMigrationRule: installSingleRule,
+    });
+
+    return (
+      <>
+        <EuiSkeletonLoading
+          isLoading={isDataLoading}
+          loadingContent={
             <>
-              <EuiFlexGroup gutterSize="m" justifyContent="flexEnd" wrap>
-                <EuiFlexItem>
-                  <SearchField initialValue={searchTerm} onSearch={handleOnSearch} />
-                </EuiFlexItem>
-                <EuiFlexItem grow={false}>
-                  <BulkActions
-                    isTableLoading={isStatsLoading || isDataLoading || isTableLoading}
-                    numberOfTranslatedRules={translationStats?.rules.installable ?? 0}
-                    numberOfSelectedRules={0}
-                    installTranslatedRule={installTranslatedRules}
-                  />
-                </EuiFlexItem>
-              </EuiFlexGroup>
-              <EuiSpacer size="m" />
-              <EuiBasicTable<RuleMigration>
-                loading={isTableLoading}
-                items={ruleMigrations}
-                pagination={pagination}
-                onChange={onTableChange}
-                selection={{
-                  selectable: () => true,
-                  onSelectionChange: setSelectedRuleMigrations,
-                  initialSelected: selectedRuleMigrations,
-                }}
-                itemId={'id'}
-                data-test-subj={'rules-translation-table'}
-                columns={rulesColumns}
-              />
+              <EuiSkeletonTitle />
+              <EuiSkeletonText />
             </>
-          )
-        }
-      />
-      {rulePreviewFlyout}
-    </>
-  );
-};
-
-export const RulesTable = React.memo(RulesTableComponent);
-RulesTable.displayName = 'RulesTable';
+          }
+          loadedContent={
+            !translationStats?.rules.total ? (
+              <EmptyMigration />
+            ) : (
+              <>
+                <EuiFlexGroup gutterSize="m" justifyContent="flexEnd" wrap>
+                  <EuiFlexItem>
+                    <SearchField initialValue={searchTerm} onSearch={handleOnSearch} />
+                  </EuiFlexItem>
+                  <EuiFlexItem grow={false}>
+                    <BulkActions
+                      isTableLoading={isStatsLoading || isDataLoading || isTableLoading}
+                      numberOfTranslatedRules={translationStats?.rules.installable ?? 0}
+                      numberOfSelectedRules={0}
+                      installTranslatedRule={installTranslatedRules}
+                    />
+                  </EuiFlexItem>
+                </EuiFlexGroup>
+                <EuiSpacer size="m" />
+                <EuiBasicTable<RuleMigration>
+                  loading={isTableLoading}
+                  items={ruleMigrations}
+                  pagination={pagination}
+                  onChange={onTableChange}
+                  selection={{
+                    selectable: () => true,
+                    onSelectionChange: setSelectedRuleMigrations,
+                    initialSelected: selectedRuleMigrations,
+                  }}
+                  itemId={'id'}
+                  data-test-subj={'rules-translation-table'}
+                  columns={rulesColumns}
+                />
+              </>
+            )
+          }
+        />
+        {rulePreviewFlyout}
+      </>
+    );
+  }
+);
+MigrationRulesTable.displayName = 'MigrationRulesTable';
