+++ x-pack/plugins/security_solution/public/siem_migrations/rules/components/rules_table_columns/name.tsx
@@ -14,14 +14,14 @@ import type { TableColumn } from './constants';
 interface NameProps {
   name: string;
   rule: RuleMigration;
-  openMigrationRulePreview: (rule: RuleMigration) => void;
+  openMigrationRuleDetails: (rule: RuleMigration) => void;
 }
 
-const Name = ({ name, rule, openMigrationRulePreview }: NameProps) => {
+const Name = ({ name, rule, openMigrationRuleDetails }: NameProps) => {
   return (
     <EuiLink
       onClick={() => {
-        openMigrationRulePreview(rule);
+        openMigrationRuleDetails(rule);
       }}
       data-test-subj="ruleName"
     >
@@ -31,15 +31,15 @@ const Name = ({ name, rule, openMigrationRulePreview }: NameProps) => {
 };
 
 export const createNameColumn = ({
-  openMigrationRulePreview,
+  openMigrationRuleDetails,
 }: {
-  openMigrationRulePreview: (rule: RuleMigration) => void;
+  openMigrationRuleDetails: (rule: RuleMigration) => void;
 }): TableColumn => {
   return {
     field: 'original_rule.title',
     name: i18n.COLUMN_NAME,
     render: (value: RuleMigration['original_rule']['title'], rule: RuleMigration) => (
-      <Name name={value} rule={rule} openMigrationRulePreview={openMigrationRulePreview} />
+      <Name name={value} rule={rule} openMigrationRuleDetails={openMigrationRuleDetails} />
     ),
     sortable: true,
     truncateText: true,
