+++ x-pack/plugins/security_solution/public/siem_migrations/rules/logic/constants.ts
