+++ x-pack/plugins/security_solution/server/lib/siem_migrations/rules/api/util/installation.ts
@@ -42,7 +42,7 @@ const installPrebuiltRules = async (
   const rulesToUpdate: UpdateRuleMigrationInput[] = [];
   const assetsToInstall: PrebuiltRuleAsset[] = [];
   rulesToInstall.forEach((ruleToInstall) => {
-    // If prebuilt rule has already been install, then just update migration rule with the installed rule id
+    // If prebuilt rule has already been installed, then just update migration rule with the installed rule id
     const installedRule = currentRules.find(
       (rule) => rule.rule_id === ruleToInstall.elastic_rule?.prebuilt_rule_id
     );
